+++ src/Config/Option/Mercure/EnableMercureConfigOption.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shlinkio\Shlink\Installer\Config\Option\Mercure;
+
+use Shlinkio\Shlink\Config\Collection\PathCollection;
+use Shlinkio\Shlink\Installer\Config\Option\AbstractSwooleDependentConfigOption;
+use Symfony\Component\Console\Style\StyleInterface;
+
+class EnableMercureConfigOption extends AbstractSwooleDependentConfigOption
+{
+    public const CONFIG_PATH = ['___', 'mercure_enabled'];
+
+    public function getConfigPath(): array
+    {
+        return self::CONFIG_PATH;
+    }
+
+    public function ask(StyleInterface $io, PathCollection $currentOptions): bool
+    {
+        return $io->confirm('Do you want to configure the integration with a Mercure hub server?', false);
+    }
+}